+++ ui/components/multichain/pages/connections/connections.js
@@ -1,34 +1,20 @@
 import React from 'react';
-import { Box, Text } from '../../../component-library';
+import { Box } from '../../../component-library';
 import {
-  AlignItems,
   BlockSize,
   Display,
   FlexDirection,
-  JustifyContent,
-  TextColor,
 } from '../../../../helpers/constants/design-system';
-import { useI18nContext } from '../../../../hooks/useI18nContext';
+import { ConnectionsHeader } from '../../connections-header';
+import { SiteNotConnected } from './components/site-not-connected';
 
-export const Connections = () => {
-  const t = useI18nContext();
-  return (
-    <Box
-      data-testid="site-not-connected"
-      display={Display.Flex}
-      flexDirection={FlexDirection.Column}
-      justifyContent={JustifyContent.center}
-      alignItems={AlignItems.center}
-      height={BlockSize.Full}
-      gap={2}
-    >
-      <Text>{t('metamaskNotConnected1')}</Text>
-      <Text color={TextColor.textAlternative}>
-        {t('metamaskNotConnected2')}{' '}
-        <Text color={TextColor.textAlternative} as="strong">
-          {t('metamaskNotConnected3')}
-        </Text>
-      </Text>
-    </Box>
-  );
-};
+export const Connections = () => (
+  <Box
+    display={Display.Flex}
+    flexDirection={FlexDirection.Column}
+    width={BlockSize.Full}
+  >
+    <ConnectionsHeader />
+    <SiteNotConnected />
+  </Box>
+);
