+++ ui/components/multichain/pages/connections/index.js
@@ -1,1 +1,2 @@
 export { Connections } from './connections';
+export { SiteNotConnected } from './components/site-not-connected';
